+++ modules/host/locals.tf
@@ -19,12 +19,56 @@ locals {
     if network_id != var.primary_network_key
   ])
 
+  server_id = coalesce(
+    try(hcloud_server.server[0].id, null),
+    try(data.hcloud_server.existing_ready[0].id, null),
+  )
+  server_name = var.existing_server_id == null ? hcloud_server.server[0].name : local.name
+  server_ipv4_address = var.existing_server_id == null ? hcloud_server.server[0].ipv4_address : try(
+    data.hcloud_server.existing_ready[0].ipv4_address,
+    null,
+  )
+  server_ipv6_address = var.existing_server_id == null ? hcloud_server.server[0].ipv6_address : try(
+    data.hcloud_server.existing_ready[0].ipv6_address,
+    null,
+  )
+  server_networks = var.existing_server_id == null ? hcloud_server.server[0].network : try(
+    data.hcloud_server.existing_ready[0].network,
+    [],
+  )
+
+  existing_server_desired_state = {
+    server_id          = var.existing_server_id
+    name               = local.name
+    location           = var.location
+    server_type        = var.server_type
+    image_id           = var.os_snapshot_id
+    labels             = merge(var.labels, { "kube-hetzner-adoption" = random_uuid.existing_server_adoption[0].result })
+    backups            = var.backups
+    firewall_ids       = sort(tolist(local.effective_firewall_ids))
+    placement_group_id = var.placement_group_id
+    networks = concat(
+      [{ id = var.network_id, ip = var.private_ipv4 }],
+      [for network_id in sort(tolist(local.extra_network_ids)) : { id = network_id, ip = null }],
+    )
+    public = {
+      ipv4 = {
+        enabled       = !var.disable_ipv4
+        primary_ip_id = var.primary_ipv4_id
+      }
+      ipv6 = {
+        enabled       = !var.disable_ipv6
+        primary_ip_id = var.primary_ipv6_id
+      }
+    }
+  }
+
   default_connection_host = coalesce(
-    hcloud_server.server.ipv4_address,
-    hcloud_server.server.ipv6_address,
+    local.server_ipv4_address,
+    local.server_ipv6_address,
     try(
-      [for network in hcloud_server.server.network : network.ip if var.network_id != null && network.network_id == var.network_id][0],
-      try([for network in hcloud_server.server.network : network.ip][0], null)
+      [for network in local.server_networks : network.ip if var.network_id != null && network.network_id == var.network_id][0],
+      try([for network in local.server_networks : network.ip][0], null)
     )
   )