+++ modules/host/out.tf
@@ -1,25 +1,25 @@
 output "ipv4_address" {
-  value      = hcloud_server.server.ipv4_address
+  value      = local.server_ipv4_address
   depends_on = [terraform_data.initial_readiness, terraform_data.os_upgrade_timer]
 }
 
 output "ipv6_address" {
-  value      = hcloud_server.server.ipv6_address
+  value      = local.server_ipv6_address
   depends_on = [terraform_data.initial_readiness, terraform_data.os_upgrade_timer]
 }
 
 output "private_ipv4_address" {
-  value      = try(one(hcloud_server.server.network).ip, "")
+  value      = try([for network in local.server_networks : network.ip if network.network_id == var.network_id][0], "")
   depends_on = [terraform_data.initial_readiness, terraform_data.os_upgrade_timer]
 }
 
 output "name" {
-  value      = hcloud_server.server.name
+  value      = local.server_name
   depends_on = [terraform_data.initial_readiness, terraform_data.os_upgrade_timer]
 }
 
 output "id" {
-  value      = hcloud_server.server.id
+  value      = local.server_id
   depends_on = [terraform_data.initial_readiness, terraform_data.os_upgrade_timer]
 }