+++ modules/host/versions.tf
@@ -12,5 +12,9 @@ terraform {
       source  = "hashicorp/cloudinit"
       version = ">= 2.3.7"
     }
+    local = {
+      source  = "hashicorp/local"
+      version = ">= 2.5.2"
+    }
   }
 }